fixup(v1): move meter delete modal
