+++ src/app/v1/facility/data/meters/meters-dashboard/meters-grouping-view/confirm-delete-group-modal/confirm-delete-group-modal.component.html
