+++ src/app/v1/facility/data/meters/meters-dashboard/meters-grouping-view/confirm-delete-group-modal/confirm-delete-group-modal.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { group } from '../../facility-meters.testing';
+import { group } from '../../../facility-meters.testing';
 import { ConfirmDeleteGroupModalComponent } from './confirm-delete-group-modal.component';
 
 describe('ConfirmDeleteGroupModalComponent', () => {
