+++ src/app/v1/facility/data/meters/meters-dashboard/meters-grouping-view/meters-grouping-view.component.ts
@@ -13,11 +13,11 @@ import {
   meterGroupTargetFromSection
 } from '../../facility-meters.models';
 import { FacilityMetersWorkspaceService } from '../../facility-meters-workspace.service';
-import { ConfirmDeleteGroupModalComponent } from '../confirm-delete-group-modal/confirm-delete-group-modal.component';
 import { MeterGroupDraftSlideoutComponent } from '../meter-dashboard-slideout/meter-group-draft-slideout/meter-group-draft-slideout.component';
 import { MeterGroupLaneComponent } from './meter-group-lane/meter-group-lane.component';
 import { MetersDashboardActionsService } from '../meters-dashboard-actions.service';
 import { MoveMeterSlideoutComponent } from '../meter-dashboard-slideout/move-meter-slideout/move-meter-slideout.component';
+import { ConfirmDeleteGroupModalComponent } from './confirm-delete-group-modal/confirm-delete-group-modal.component';
 
 @Component({
   selector: 'app-meters-grouping-view',
